Module Swscale

This module perform image scaling and color space/pixel format conversion operations.

val version : unit -> int

val configuration : unit -> string

val license : unit -> string

type pixel_format = Avutil.Pixel_format.t

type flag =

| Fast_bilinear

| Bilinear

| Bicubic

| Print_info

type t

val create : flag list -> int -> int -> pixel_format -> int -> int -> pixel_format -> t

Swscale.create flags in_w in_h in_pf out_w out_h out_pf create a Swscale.t scale context with the flags flag list defining the conversion type, the in_w width, in_h height and in_pf pixel format for input format and out_w width, out_h height and out_pf pixel format for output format.

type planes = (Avutil.data * int) array

val scale : t -> planes -> int -> int -> planes -> int -> unit

Swscale.scale ctx in_planes y h out_planes off scale the h rows of in_planes strarting from the row y to the off row of the out_planes output.

module type VideoData = sig ... end

Video data modules for Swscale module input and output parameterization.

type ('i, 'o) ctx

module Make : functor (I : VideoData) -> functor (O : VideoData) -> sig ... end

Functor building an implementation of the swscale structure with parameterized input an output video data types

module BigArray : sig ... end

Unsigned 8 bit bigarray.

module Frame : sig ... end

Video frame.

module Bytes : sig ... end

Bytes array.