Module Swscale.Make

Functor building an implementation of the swscale structure with parameterized input an output video data types

Parameters

I : VideoData

O : VideoData

Signature

type t = (I.t, O.t) ctx

val create : flag list -> int -> int -> pixel_format -> int -> int -> pixel_format -> t

Swscale.create flags in_w in_h in_pf out_w out_h out_pf do the same as Swscale.create.

val convert : t -> I.t -> O.t

Swscale.convert ctx ivd scale and convert the ivd input video data to the output video data according to the ctx scaler context format.

Raise Error if the conversion failed.